Move waitUntilScopeAndCollectionExists over
